requirements.txt
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
beautifulsoup4==4.13.5
blinker==1.9.0
brotlipy==0.7.0
bs4==0.0.2
bump2version==1.0.1
bumpversion==0.6.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
docutils==0.22
extruct==0.18.0
filelock==3.19.1
Flask==3.1.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
html5lib==1.1
html_text==0.7.0
httplib2==0.30.0
hyperlink==21.0.0
id==1.5.0
idna==3.10
incremental==24.7.2
isort==6.0.1
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jstyleson==0.0.2
keyring==25.6.0
loguru==0.7.3
lxml==6.0.1
lxml_html_clean==0.4.2
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mf2py==2.0.1
more-itertools==10.8.0
msgpack==1.1.1
nh3==0.3.0
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
parsel==1.10.0
pdoc3==0.11.6
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyRdfa3==3.6.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
queuelib==1.8.0
rdflib==7.1.4
readme_renderer==44.0
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rsa==4.9.1
scrapfly-sdk==0.8.23
Scrapy==2.13.3
SecretStorage==3.3.3
sentry-sdk==2.35.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
tldextract==5.3.0
twine==6.1.0
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
zope.interface==7.2
